requirements.txt
flask==3.0.3
flask-assets==2.1.0
libsass==0.23.0
mysql-connector==2.2.9